ESCKAN-65 move doc definition to pdf service
